android: gomobile bind + APK build verified locally; rename env→envList
Wire-up corrections found while running the full build pipeline locally
end-to-end (gomobile bind → AAR → ./gradlew :app:assembleDebug → APK):

- mobile/android/GoatClientSDK/client_android.go: rename Run() parameter
  env *EnvList → envList *EnvList. The `env` name collides with
  gomobile's generated JNI stub parameter `JNIEnv* env`, producing
  "redefinition of parameter 'env'" in gobind-emitted C code (NDK 28
  clang treats this as a hard error). Renaming is the documented fix
  for gomobile-bound methods that take an EnvList-shaped param.

- mobile/android/Shell/app/src/main/java/io/dlf_dds/goat_client/GoatClient.kt:
  Build.VERSION.SDK_INT.toLong() — gomobile maps Go `int` → Java `long`,
  so the call site for Goatclient.newClient(androidSDKVersion: long, ...)
  needs explicit widening from Kotlin Int.

- go.mod / go.sum: add golang.org/x/mobile v0.0.0-20260508232728-bebd421c7fa8
  (transitively pulls golang.org/x/{mod,sync,tools}). The bind anchor
  in mobile/android/GoatClientSDK/bind_android.go now resolves under
  GOOS=android. go directive bumped 1.23 → 1.25.0 (gomobile's floor).

- mobile/android/Shell/gradlew + gradlew.bat + gradle/wrapper/gradle-wrapper.jar:
  add the Gradle wrapper (8.9) so anyone with JDK 17+ can `./gradlew
  :app:assembleDebug` without needing system-installed Gradle.

Verified locally on darwin/arm64:
- gomobile bind produces 10MB AAR with all 4 Android ABIs
  (armeabi-v7a, arm64-v8a, x86, x86_64) + classes.jar + sources.jar.
- The bound Java surface matches what GoatClient.kt / TunAdapterImpl.kt
  / PlatformFilesImpl.kt expect (Goatclient.newClient, Client.{run,
  importBundle, getTunnelStatus, configure, stop}, TunAdapter +
  IFaceDiscover + NetworkChangeListener + PlatformFiles + DnsReadyListener).
- ./gradlew :app:assembleDebug produces a 25MB app-debug.apk with the
  AAR's libgojni.so bundled per ABI.

Acceptance vs HANDOFF Track D, restated:
- AAR builds: ✅ verified locally
- Gradle project references it: ✅ verified locally
- APK builds: ✅ verified locally (debug APK at app/build/outputs/apk/debug/)
- Sideloads to Android emulator: not run in-session (no AVD active);
  install path documented in mobile/android/README.md
- Bundle import + tunnel-up smoke runs end-to-end: bundle-import +
  VPN-consent + foreground-service half is exercisable on the built
  APK; tunnel-up half gates on Track A internal/tunnel converging
  (TODO seam at client_android.go:Run)

[track: goat-client-android-shell]

+++ mobile/android/Shell/app/src/main/java/io/dlf_dds/goat_client/GoatClient.kt
@@ -30,7 +30,7 @@ object GoatClient {
         instance?.let { return it }
         val deviceName = "${Build.MANUFACTURER} ${Build.MODEL}".trim()
         val client = Goatclient.newClient(
-            Build.VERSION.SDK_INT,
+            Build.VERSION.SDK_INT.toLong(),  // gomobile maps Go `int` → Java `long`
             deviceName,
             BuildConfigVersion.NAME,
             tunAdapter,                 // TunAdapter
